# Fix fuzz 1 failure: incorrect reduction of BigInt