From the Ethereum Foundation sponsored fuzzing by @guidovranken (linked to #54),
``fromBig`` can lead to incorrectly reduced field elements.

This is because classic Montgomery reduction can reduce inputs from the range [0, 4p²) -> [0, p)  but this is not the case with the no-carry optimization from Gnark (https://hackmd.io/@gnark/modular_multiplication and https://eprint.iacr.org/2022/1400.pdf section 2)

In practice ``getMont`` and ``fromBig`` are internal-only procedures that only deserialized well-formed inputs that have been pre-checked in protocols:
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/ethereum_bls_signatures.nim#L284-L293
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/ethereum_bls_signatures.nim#L339-L354
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/ethereum_evm_precompiles.nim#L38-L54

and for large inputs, we use redc2xMont + multMont:
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/hash_to_curve/h2c_hash_to_field.nim#L222-L235
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/ethereum_eip2333_bls12381_key_derivation.nim#L56-L71

or full reduction:
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/math_arbitrary_precision/arithmetic/bigints_views.nim#L69-L79

However there is one protocol case where ``getMont`` is used, in ECMUL (EIP-198):
- https://github.com/mratsim/constantine/blob/151f284da6caad1545a38f871c421c5bb2da8f6a/constantine/ethereum_evm_precompiles.nim#L189-L209

## Perf

The impact on BN254 is a slowdown from 10ns to 17ns (with Clang, worse with GCC)

The impact on BLS12-381 is a slowdown from 23ns to 39ns (with Clang, worse with GCC)

Most of the slowdown is due to classic Montgomery reduction not having an assembly implementation.

However, that conversion is not in the hot path of any protocol, most protocols that need to deal with a large amount of deserialization allow caching, for example in case of Ethereum validator public keys. And 99% of the time is spend in computing the square root of compressed keys/signatures.